Extract `SongSearch` internal classes
Introduces `SongSearch::Document` to parse the response HTML. In the
future, we'll be able to handle more edge cases by testing
`SongSearch::Document` parsing more closely.

Introduces `SongSearch::Request` to handle making requests to
SkateVideoSite. It accepts a `SongSearch` instance and a `cache:`
argument, which it expects to be `Hash`-like.

From the `IdentificationsController`, we construct the `SongSearch` with
an instance of `Rails.cache`, which it passes along to the
`SongSearch::Request`.