Simplify `identifications` controller
Style the `[data-target="identifications.recordButton"]` element to make
it easier to tap on smaller devices.

Once recording, set `<button disabled>` and add an appropriate class to
the controller element. Once results have been published, removed the
class and enable the `<button>`.

Remove the `<ul>` from the page, since we'll be redirecting once we get
results.