Translate Stimulus code to TypeScript
Now that our environment has configured `webpacker` to support
TypeScript, this commit translates existing code from JavaScript to
TypeScript.

Some interesting details to highlight:

1. Our `recordings` controller now specifies each of the targets as the
_most specific_ type of [`Element`][Element] descendant as possible
(e.g. [`HTMLInputElement`][HTMLInputElement],
[`HTMLButtonElement`][HTMLButtonElement]). This decision hinges on the
`data-target` annotations being attached to the proper Nodes!

2. Our `bridge` module has been modified to export a factory of sorts,
instead of an instance. Now that we have types, the module can export
the `Bridge` interface to its consumers.

When the `window.webkit` variable is unavailable to us (for instance,
when we're developing on a device that is NOT running `iOS`), return a
`FakeBridge` instance which implements the `Bridge` interface.

When the `window.webkit` variable is available to us, dig into it to
return the `GripThumb` message handler.

3. Extract a private function to communicate the we _must submit the
`<form>` element with a button click_, so that the corresponding
[`submit` event][submit] is fired:

> The `submit` event only fires when the user clicks a submit button
> (`<button>` or `<input type="submit">`) in a form. The event is not
> raised when calling the `form.submit()` method directly.

The `submit` event is necessary for Rails-UJS compliant,
`data-remote="true"` forms to submit and integrate with Turbolinks.

[Element]: https://developer.mozilla.org/en-US/docs/Web/API/Element
[HTMLInputElement]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLInputElement
[HTMLButtonElement]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLButtonElement
[submit]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLFormElement/submit_event